Add more types for typedoc
