+++ packages/viewer/src/components/scene-tree-table-layout/scene-tree-table-layout.tsx
@@ -18,10 +18,10 @@ import {
 } from '@vertexvis/html-templates';
 
 import { readDOM } from '../../lib/stencil';
-import { RowDataProvider } from '../scene-tree/interfaces';
 import { SceneTreeController } from '../scene-tree/lib/controller';
 import { getSceneTreeViewportHeight } from '../scene-tree/lib/dom';
 import { isLoadedRow, LoadedRow, Row } from '../scene-tree/lib/row';
+import { RowDataProvider } from '../scene-tree/types';
 import {
   DomScrollToOptions,
   getSceneTreeTableOffsetTop,
