+++ packages/viewer/src/components/scene-tree/lib/row.ts
@@ -1,6 +1,6 @@
 import { Node } from '@vertexvis/scene-tree-protos/scenetree/protos/domain_pb';
 
-import { MetadataKey } from '../interfaces';
+import { MetadataKey } from '../types';
 
 export type MetadataMap = Record<MetadataKey, string | undefined>;
