+++ packages/viewer/src/components/viewer-box-query-tool/types.ts
@@ -1,0 +1,6 @@
+export type VolumeIntersectionQueryType =
+  | 'clearAndSelect'
+  | 'select'
+  | 'deselect';
+
+export type VolumeIntersectionQueryMode = 'exclusive' | 'inclusive';
