+++ packages/viewer/src/components/viewer-toolbar-group/types.ts
@@ -1,0 +1,1 @@
+export type ViewerToolbarGroupDirection = 'horizontal' | 'vertical';
