+++ packages/viewer/src/components/viewer-toolbar/viewer-toolbar.tsx
@@ -1,20 +1,7 @@
 import { Component, h, Prop } from '@stencil/core';
 import classname from 'classnames';
 
-/**
- * Values that define where the toolbar is positioned.
- */
-export type ViewerToolbarPlacement =
-  | 'top-left'
-  | 'top-center'
-  | 'top-right'
-  | 'middle-left'
-  | 'middle-right'
-  | 'bottom-left'
-  | 'bottom-center'
-  | 'bottom-right';
-
-export type ViewerToolbarDirection = 'horizontal' | 'vertical';
+import { ViewerToolbarDirection, ViewerToolbarPlacement } from './types';
 
 @Component({
   tag: 'vertex-viewer-toolbar',
