+++ packages/viewer/src/index.ts
@@ -6,8 +6,11 @@ import * as ColorMaterial from './lib/scenes/colorMaterial';
 export { ColorMaterial };
 
 export * from './components';
-export * from './components/scene-tree/interfaces';
 export * from './components/scene-tree/lib';
+export * from './components/scene-tree/types';
+export * from './components/viewer-box-query-tool/types';
+export * from './components/viewer-toolbar/types';
+export * from './components/viewer-toolbar-group/types';
 export { Config } from './lib/config';
 export * from './lib/cursors';
 export { Environment } from './lib/environment';
@@ -15,6 +18,7 @@ export * from './lib/formatter';
 export * from './lib/interactions/interactionHandler';
 export * from './lib/interactions/keyInteraction';
 export * from './lib/measurement';
+export * from './lib/pins';
 export { Scene } from './lib/scenes';
 export {
   EntityType,
