+++ packages/viewer/src/lib/pins/index.ts
@@ -1,0 +1,17 @@
+import {
+  IconPin,
+  Pin,
+  PinStyleAttributes,
+  TextPin,
+  ViewerPinToolMode,
+  ViewerPinToolType,
+} from './model';
+
+export {
+  IconPin,
+  Pin,
+  PinStyleAttributes,
+  TextPin,
+  ViewerPinToolMode,
+  ViewerPinToolType,
+};
